#253: Move tests to new file
