+++ src/backend/tests/projects.test.ts
@@ -2,11 +2,7 @@ import request from 'supertest';
 import express from 'express';
 import projectRouter from '../src/routes/projects.routes';
 import prisma from '../src/prisma/prisma';
-import {
-  createRulesChangesJson,
-  getChangeRequestReviewState,
-  getHighestProjectNumber
-} from '../src/utils/projects.utils';
+import { getChangeRequestReviewState, getHighestProjectNumber } from '../src/utils/projects.utils';
 import { batman } from './test-data/users.test-data';
 
 const app = express();
@@ -122,53 +118,5 @@ describe('Projects', () => {
     const res = await request(app).post('/edit').send(proj);
 
     expect(res.statusCode).toBe(400);
-  });
-
-  test('createRulesChangesJson with empty old + new lists', () => {
-    const rulesChanges = createRulesChangesJson('test', [], [], 1, 1, 1);
-    expect(rulesChanges.length).toEqual(0);
-  });
-
-  test('createRulesChangesJson with empty new list and non-empty old list', () => {
-    const rules = ['rule1', 'rule2', 'rule3'];
-    const rulesChanges = createRulesChangesJson('test', rules, [], 1, 1, 1);
-    expect(rulesChanges.length).toEqual(3);
-    rulesChanges.forEach((r, i) => {
-      expect(r.changeRequestId).toEqual(1);
-      expect(r.implementerId).toEqual(1);
-      expect(r.wbsElementId).toEqual(1);
-      expect(r.detail).toEqual(`Removed test "${rules[i]}"`);
-    });
-  });
-
-  test('createRulesChangesJson with empty old list and non-empty new list', () => {
-    const rules = ['rule1', 'rule2', 'rule3'];
-    const rulesChanges = createRulesChangesJson('test', [], rules, 1, 1, 1);
-    expect(rulesChanges.length).toEqual(3);
-    rulesChanges.forEach((r, i) => {
-      expect(r.changeRequestId).toEqual(1);
-      expect(r.implementerId).toEqual(1);
-      expect(r.wbsElementId).toEqual(1);
-      expect(r.detail).toEqual(`Added new test "${rules[i]}"`);
-    });
-  });
-
-  test('createRulesChangesJson with non-empty old list and non-empty new list', () => {
-    const oldRules = ['rule1', 'rule2', 'rule3'];
-    const newRules = ['rule4', 'rule5', 'rule6'];
-    const rulesChanges = createRulesChangesJson('test', oldRules, newRules, 1, 1, 1);
-    expect(rulesChanges.length).toEqual(6);
-    rulesChanges.slice(0, 3).forEach((r, i) => {
-      expect(r.changeRequestId).toEqual(1);
-      expect(r.implementerId).toEqual(1);
-      expect(r.wbsElementId).toEqual(1);
-      expect(r.detail).toEqual(`Removed test "${oldRules[i]}"`);
-    });
-    rulesChanges.slice(3).forEach((r, i) => {
-      expect(r.changeRequestId).toEqual(1);
-      expect(r.implementerId).toEqual(1);
-      expect(r.wbsElementId).toEqual(1);
-      expect(r.detail).toEqual(`Added new test "${newRules[i]}"`);
-    });
   });
 });
